chore: split input module in CM4 in definitions and API #7
